Hide start menu replacement option on Windows 11
